requirements.txt
python-dotenv
setuptools
pydantic
numpy
uvicorn
mcp
fastapi
fastapi-mcp
langchain
langchain-text-splitters
langchain-community
langchain-ollama
langchain-milvus
langchain-mongodb
langchain-neo4j
llama-index
llama-index-core
llama-index-vector-stores-milvus
llama-index-embeddings-ollama
llama-index-llms-ollama
llama-index-embeddings-langchain
llama-index-tools-neo4j
llama-index-graph-stores-neo4j
llama-index-packs-neo4j-query-engine
transformers
sentence-transformers
ollama
pymilvus
pymongo
redis
httpx
aiohttp
tiktoken
torch
sqlalchemy
pymssql
jinja2
pika
neo4j
numpy
pandas
yomitoku
Pillow
PyMuPDF
opencv-python
pytesseract
ebooklib
yomitoku
markdown
chardet